chore: resolve review remarks cy-21
