+++ apps/frontend/src/assets/css/reset.css
@@ -24,7 +24,6 @@ body {
 	height: 100%;
 	padding: 0;
 	margin: 0;
-	overflow: hidden;
 }
 
 body {
