+++ apps/frontend/src/pages/home/components/feedbacks-section/components/feedback-card/styles.module.css
@@ -21,6 +21,7 @@
 	overflow: hidden;
 	text-overflow: ellipsis;
 	-webkit-line-clamp: 5;
+	line-clamp: 5;
 	-webkit-box-orient: vertical;
 }
