+++ apps/frontend/src/pages/home/components/feedbacks-section/components/user-avatar/user-avatar.tsx
@@ -2,12 +2,12 @@ import { type FC } from "react";
 
 import styles from "./styles.module.css";
 
-interface Properties {
+type AvatarProperties = {
 	alt: string;
 	src: string;
 }
 
-const UserAvatar: FC<Properties> = ({ alt, src }) => {
+const UserAvatar: FC<AvatarProperties> = ({ alt, src }) => {
 	return (
 		<div className={styles["wrapper"]}>
 			<img alt={alt} className={styles["image"]} src={src} />
