+++ apps/frontend/src/pages/home/lib/constants.ts
@@ -1,0 +1,22 @@
+const FEEDBACKS = [
+  {
+    avatar: "src/assets/img/home/roy.jpg",
+    id: 1,
+    name: "Roy",
+    text: "Lorem ipsum dolor amet, consectetur adipiscing elit. Cras sed dui sagittis, scelerisque lectus at, porttitor lectus. Sed libero est, tincidunt eget purus nec, dignissim consequat mauris",
+  },
+  {
+    avatar: "src/assets/img/home/emma.jpg",
+    id: 2,
+    name: "Emma",
+    text: "Nulla et nulla pulvinar, congue justo id, cursus ligula. Nunc pharetra sapien libero, vel blandit orci rhoncus ut. Sed aliquam efficitur semper.",
+  },
+  {
+    avatar: "src/assets/img/home/joan.jpg",
+    id: 3,
+    name: "Joan",
+    text: "Nullam tempus, elit non tempus molestie, tellus diam sagittis urna, vel viverra velit risus in nunc. Cras in quam leo. Nullam mattis at lacus eget pretium. Etiam quis pulvinar",
+  },
+] as const;
+
+export { FEEDBACKS };
