+++ apps/frontend/src/pages/home/styles.module.css
@@ -1,5 +1,5 @@
 main {
-	height: 100vh;
+	min-height: 100vh;
 	overflow-y: auto;
 	scroll-snap-type: y mandatory;
 	scrollbar-width: none; /* Firefox */
@@ -14,7 +14,6 @@ section {
 	display: flex;
 	align-items: center;
 	justify-content: center;
-	height: 100vh;
 	min-height: 100vh;
 	scroll-snap-align: start;
 }
